Pokemon -- Cleaned Code
I cleaned up the code
Inside the init of the pokemon I added a power mode which checks the number of stars the pokemon has for attacks!

I put lots of things into functions to clean it up

Most importantly I made a new option called (Catch) where the user has the ability to try and capture the pokemon it has encountered... I will be making the class for pokedex soon!